requirements.txt
asgiref==3.3.4
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==3.4.7
cssselect==1.1.0
Django==3.2
django-crispy-forms==1.11.2
django-phonenumber-field==5.1.0
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
parsel==1.6.0
phonenumbers==8.12.21
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pytz==2021.1
queuelib==1.5.0
requests==2.25.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.7.4.3
urllib3==1.26.4
w3lib==1.22.0
webdriver-manager==3.4.0
whitenoise==5.2.0
zope.interface==5.4.0
